ducky: introduce compaction_e2e_test
compaction e2e test is an iteration of compaction end to end test
which switches to more powerful verifier for idempotency workload
which also isn't affected by the ssh errors
